Speech_To_Text (Offline Using Openai's Whisper)/Reatime/Method 2/requirement.txt
openai-whisper==20240930
PyAudio==0.2.14
ffmpeg
